+++ CHANGES.rst
@@ -28,6 +28,8 @@ Features Added
 Restructuring
 +++++++++++++
 
+- Drop `OFS.History` functionality.
+
 - Removed ZMI export/import feature.
 
 - Drop ZopeUndo dependency and move undo management to the control panel.